Fixed up more of the factory landing page
Made the status colour to be the background of the label instead of just the status text, for better visibility and for consistency with the new machines page.

Made use of the assigned operator field in machines which has since been added since my previous pull request. If there is no one assigned, it states it as Vacant.

Archived machines are not added to the factory landing page.

Weird issue with the first line '3D Printer' seeming to not follow the 'text-align: center' property, but all the rest do. Not sure if you guys know a quick fix for that, as all I came across was that it could potentially be an issue with tables? But not sure why that would still be happening only to the first element.

So next would be to make each item clickable to re-direct to that machine's individual page for factory logs?