Changing the factory page based on Arran's suggestions
Factory page is now displaying factory log data, and not its initial redundant parameters such as status, assigned PO, and location (which are all also displayed on the machines page).

The page still has the colour indicator for the machine status. The 7 quantitative fields from the Log data for each machine are displayed as they are in their current state (a.k.a. at the end of the csv). I have tried to order them with what a manager, or even the POs, would care for the most towards the top, with possible fields that could act as warning signs toward the bottom.

However, some of my code is cheating a little bit as I am getting quite confused about how similar functions have been implemented on the other pages. My SQL query for fetching the Log data is manually writing in the most up-to-date timestamp, instead of finding that out properly. I am also manually associating the names of the machines with their machineID, as I keep trying to implement the same getMachineName() function from the Reports page, but it keeps breaking and I'm not entirely sure where I am going wrong there. I also still haven't properly implemented the buttons at the bottom. I'd like to use the same date input from the Reports page as well, but I am getting very lost looking into that page. The +/- 30min buttons should also be implemented by adjusting the SQL query?

Apologies for not entirely finishing this page before the pull request, but I think I do need some help with the final pieces. I normally would like to bog down for a week and really research this properly, but especially with the presentation due so soon, I am hoping a little help from you guys could speed me up quite a lot.

Let me know if there are any styling issues with the spacing, last time seemed to be a bit sensitive and did not look the same as on my screen. There should be 3 columns.